Added file extensions
Added GoPro and Lumix FZ1000 image types to $MEDIA_REGEX.

Next, I need to look at where the code is to remove the image files from the SD cards. Maybe need to add a flag (file) which can be placed on the SD card to indicate DCIM needs to be wiped after a successful rsync.