Source options & checks
A camera of mine puts videos in a separate path outside of DCIM. Placing a file (sources.cnf) on the backup target in the .vst folder, containing one or more paths (one per line). Will cause these paths to be uses as sources in addition to DCIM.

Added checks to test for empty folders, no point running rsync if there's nothing to do.